Next assignment (Scala & Haskell version).
The Haskell version, while shorter, is much more complicated,
beacuse the dictionary is read from file, and is thus stateful.
The implementation uses the Reader monad to pass the words in the background,
which makes sentenceAnagrams a bit convoluted.

It would have been simpler to just
 i) pass the dictionary to anagramsByOccurrences, wordAnagrams and sentenceAnagrams
ii) use a static dictionary (a Haskell source file)
... but I though this is the `correct' solution.